apps/fastapi/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = "FastAPI for processing audio files into transcripts and SOAP notes."
authors = ["trevorpfiz <elektrikspark@gmail.com>", "zacharypfiz <ztpfizcode@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
uvicorn = "^0.30.1"
pydantic-settings = "^2.3.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
requests = "^2.32.3"
fastapi = "^0.111.0"
openai = "^1.35.6"
# Therapy module dependencies
torch = "^2.1.0"
transformers = "^4.36.0"
librosa = "^0.10.1"
numpy = "^1.26.0"
scipy = "^1.11.0"
soundfile = "^0.12.1"

[tool.poetry.group.therapy.dependencies]
# Optional: heavier ML models (install with: poetry install --with therapy)
speechbrain = "^1.0.0"
whisperspeech = {version = "^0.8.0", optional = true}

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.0"
pytest = "^8.2.2"
pylint-pydantic = "^0.3.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
